pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "genset"
version = "4.0.0"
description = "Generator and Booking Management System"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14.1,<2",
    "pandas>=1.5.3",
    "openpyxl>=3.1.0",
    "xlrd>=2.0.1",
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "jinja2==3.1.3",
    "python-multipart==0.0.6",
    "passlib[bcrypt]==1.7.4",
    "itsdangerous==2.2.0",
    "bcrypt==3.2.2",
    "PyJWT==2.8.0",
    "psutil>=5.9.0",
    "psycopg[binary]>=3.2.6,<4",
    "psycopg_pool>=3.2.6,<4",
    "python-dotenv>=1.0.1,<2",
    "slowapi>=0.1.9,<1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.27.0,<0.28.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v"

[tool.setuptools]
packages = ["core", "cli", "web", "scripts"]
